requirements.txt
streamlit
PyPDF2
python-docx
transformers
torch
sentencepiece
accelerate